version 1.2

+++ assets/js/script.js
@@ -4,10 +4,10 @@ console.log(celsius, fahrenheit);
 
 
 function celToFar(){
-    let output = ( parseFloat(celsius.value) * 9/5) + 32;
-    fahrenheit.value = parseFloat(output.toFixed(2));
+    let output = ( parseInt(celsius.value) * 9/5) + 32;
+    fahrenheit.value = parseInt(output.toFixed(2));
 }
 function farToCel(){
-    let output = ( parseFloat(fahrenheit.value) - 32) * 5/9;
-    celsius.value = parseFloat(output.toFixed(2));
+    let output = ( parseInt(fahrenheit.value) - 32) * 5/9;
+    celsius.value = parseInt(output.toFixed(2));
 }
